fix(BaseballForm) : rename variables
